rename the tab1.tsx into profile.tsx and also the tab1 code
